Fix harmless compiler warnings reported by MSVC.
FossilOrigin-Name: 391e21bb6e0b2f4632972b6617a3a18192f88deb29eee5bc060846468e624b21

+++ src/vdbeapi.c
@@ -1864,7 +1864,7 @@ int sqlite3_stmt_explain(sqlite3_stmt *pStmt, int eMode){
   Vdbe *v = (Vdbe*)pStmt;
   int rc;
   sqlite3_mutex_enter(v->db->mutex);
-  if( v->explain==eMode ){
+  if( ((int)v->explain)==eMode ){
     rc = SQLITE_OK;
   }else if( eMode<0 || eMode>2 ){
     rc = SQLITE_ERROR;
